requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.11.1
anyio==3.7.0
attrs==24.2.0
bcrypt==4.0.1
bidict==0.23.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.97.0
fastapi-socketio==0.0.10
frozenlist==1.4.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.5
orjson==3.9.1
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pytest==7.3.2
python-dotenv==1.0.0
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.11.3
PyYAML==6.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.4
